+++ src/test/scala/edu/ie3/mobsim/model/ChargingBehaviorTestData.scala
@@ -47,23 +47,21 @@ trait ChargingBehaviorTestData extends TripSimulationTestData {
     departureTime = ZonedDateTime.now().plusHours(5)
   )
 
-  protected val currentPricesAtChargingStations: Map[UUID, java.lang.Double] = {
+  protected val currentPricesAtChargingStations: Map[UUID, Double] = {
     chargingStations.map { chargingStations =>
-      chargingStations.uuid -> java.lang.Double.valueOf(0.0)
+      chargingStations.uuid -> 0.0
     }.toMap
   }
 
-  protected val currentlyAvailableChargingPoints: Map[UUID, Integer] = {
+  protected val currentlyAvailableChargingPoints: Map[UUID, Int] = {
     chargingStations.map { chargingStations =>
-      chargingStations.uuid -> Integer.valueOf(
-        chargingStations.chargingPoints
-      )
+      chargingStations.uuid -> chargingStations.chargingPoints
     }.toMap
   }
 
-  protected val noAvailableChargingPoints: Map[UUID, Integer] = {
+  protected val noAvailableChargingPoints: Map[UUID, Int] = {
     chargingStations.map { chargingStations =>
-      chargingStations.uuid -> Integer.valueOf(0)
+      chargingStations.uuid -> 0
     }.toMap
   }
